refactor(cli): simplify Shift+Enter to core keypress detection only
Modern terminals (VSCode, Cursor, iTerm2, Ghostty, Kitty, WezTerm, Warp)
already support Shift+Enter natively via the Kitty keyboard protocol.
The /terminal-setup command and native macOS modifier detection were
over-engineered solutions that either didn't work reliably or were
unnecessary for the target use cases.

Remove:
- packages/modifiers-napi: native C++ addon for Apple Terminal modifier
  detection — unreliable due to race condition between Shift key release
  and Node.js event processing
- terminalSetup.ts + terminalSetupCommand.ts: the /terminal-setup command
  and all terminal-specific configuration (VSCode keybindings, Alacritty
  TOML, Zed keymap, Apple Terminal PlistBuddy)
- All related i18n keys, tests, and build config references

Keep (the only code actually needed):
- Kitty protocol CSI-u sequence parsing (handles most modern terminals)
- ESC+CR (meta=true) detection for terminals sending \x1b\r
- Backslash+Enter fallback for any terminal
- Ctrl+J universal newline (already in keyMatchers, no change needed)

Rename shift-enter-newline.test.ts → newline-insertion.test.ts,
keeping only the key binding matching and backslash+Enter tests.

E2E: 5/5 scenarios pass (plain Enter, Ctrl+J, ESC+CR, Kitty CSI-u,
backslash+Enter).
