--- packages/modifiers-napi/.gitignore
@@ -1,3 +1,0 @@
-build/
-prebuilds/
-node_modules/
